requirements.txt
asgiref= 3.5.2 
certifi= 2022.9.24 
crispy-bootstrap5-0.7 
distlib=-0.3.6 
Django -4.1.2 
django-cors-headers==3.13.0 
django-crispy-forms=1.14.0 
djangorestframework 3.14.0 
filelock--3.8.0 
pipenv=-2022.10.25 
platformdirs==2.5.2 
pytz==2022.5 
sqlparse==0.4.3 
tzdata-2022.5 
virtualenv=-20.16.6 
virtualenv-clone==0.5.7